+++ ldaptor/_encoder.py
@@ -34,6 +34,16 @@ def to_unicode(value):
     return value
 
 
+def repr_converter(value):
+    """
+    Converts value to its string representation:
+
+    * Byte string for Python 2
+    * Unicode string for Python 3
+    """
+    return to_bytes(value) if six.PY2 else to_unicode(value)
+
+
 def get_strings(value):
     """
     Getting tuple of available string values